Renames folders and adds explanation of this to top of readme
