+++ hapiedge/ch4-handler/lib/controllers/welcome.js
