+++ hapiedge/ch4-handler/package.json
@@ -4,6 +4,7 @@
   "description": "Working through 'Developing a Hapi Edge' book",
   "main": "index.js",
   "scripts": {
+    "start": "node index.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "repository": {
@@ -16,14 +17,12 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "async": "^1.4.2",
     "bell": "^5.3.0",
     "blipp": "^2.3.0",
     "glue": "^2.3.0",
     "good": "^6.4.0",
     "good-console": "^5.1.0",
     "hapi": "^10.1.0",
-    "hapi-auth-cookie": "^3.1.0",
-    "hoek": "^2.16.3"
+    "hapi-auth-cookie": "^3.1.0"
   }
 }
