+++ hapiedge/ch5-validation/.gitignore
@@ -1,0 +1,7 @@
+npm-debug.log
+node_modules
+
+#Mac DS_Store file
+.DS_Store
+*/.DS_Store
+*/*/.DS_Store
